requirements.txt
flask==3.0.3
flask-httpauth==4.8.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.3
anthropic==0.28.0
python-dotenv==1.0.1
apscheduler==3.10.4
gunicorn==22.0.0
